@@ -5,6 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "nodemon server/index.js",
+    "build": "webpack -w",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -32,14 +33,21 @@
     "babel-preset-react": "^6.24.1",
     "css-loader": "^6.7.1",
     "html-webpack-plugin": "^5.5.0",
+    "sass": "^1.54.8",
+    "sass-loader": "^13.0.2",
     "source-map-loader": "^1.0.0",
     "style-loader": "^3.3.1",
     "uglify-js": "^3.15.4",
-    "webpack": "^5.70.0",
+    "webpack": "^5.74.0",
     "webpack-cli": "^4.9.2",
     "webpack-dev-server": "^4.7.4"
   },
   "dependencies": {
+    "@emotion/react": "^11.10.4",
+    "@emotion/styled": "^11.10.4",
+    "@fortawesome/fontawesome-svg-core": "^6.2.0",
+    "@fortawesome/free-solid-svg-icons": "^6.2.0",
+    "@fortawesome/react-fontawesome": "^0.2.0",
     "@material-ui/core": "^4.12.3",
     "@material-ui/styles": "^4.11.4",
     "@mui/icons-material": "^5.5.1",
@@ -65,12 +73,16 @@
     "jwthelper": "^0.0.4",
     "lodash": "^4.17.21",
     "mongoose": "^6.2.7",
+    "node-sass": "^7.0.1",
     "nodemon": "^2.0.15",
     "npm": "^8.19.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-hook-form": "^7.29.0",
+    "react-icons": "^4.4.0",
+    "react-owl-carousel": "^2.3.3",
     "react-router-dom": "^6.3.0",
+    "react-transition-group": "^4.4.5",
     "sib-api-v3-sdk": "^8.4.2",
     "styled-components": "^5.3.5",
     "url-loader": "^4.1.1",
