@@ -1,99 +1,43 @@
-import React, { useEffect, useState } from 'react';
-import axios from 'axios';
-import { useNavigate } from 'react-router-dom';
-import Card from '@mui/material/Card';
-import Box from '@mui/material/Box';
-import Button from '@mui/material/Button';
-import { useForm } from "react-hook-form";
-import IconButton from '@mui/material/IconButton';
-import Alert from '@mui/material/Alert';
-import Typography from '@mui/material/Typography';
-import DialogActions from '@mui/material/DialogActions';
-import Modal from '@mui/material/Modal';
-import PersonIcon from '@mui/icons-material/Person';
-import CheckIcon from '@mui/icons-material/Check';
-import { useMainContext } from '../../Providers/MainProvider.jsx';
-import { Dialog, DialogContent } from "@material-ui/core";
-import { makeStyles } from "@material-ui/styles";
+import React, { useEffect, useState } from 'react'
+import axios from 'axios'
+import { useMainContext } from '../../Providers/MainProvider.jsx'
+import { useNavigate } from 'react-router-dom'
+import Typography from '@mui/material/Typography'
+import Button from '@mui/material/Button'
+import NavigationBar from '../../Layout/WithNavigation.jsx'
+import TextField from '@mui/material/TextField'
+import Grid from '@mui/material/Grid'
+import Box from '@mui/material/Box'
+import '../../Styles.scss'
 
+import { useLayoutEffect } from 'react'
 
-
-
-const swipeStyle = {
-  position: "absolute",
-  display: "inline-block",
-  zIndex: 9998
-}
-
-const cardStyle = {
-  position: "absolute",
-
-  backgroundColor: "#ff9800",
-  width: 330,
-  maxWidth: "85vw",
-  height: "60vh",
-  borderRadius: '10px',
-  'backgroundSize': "cover",
-  boxShadow: '0px 0px 5px 0px rgba(0,0,0,0.05)',
-  'backgroundPosition': "center"
-}
-
-const cardContainer = {
-  textAlign:'center',
-  width: '90vw',
-  maxWidth: '330px',
-  height: '70vh',
-  marginTop: '5vh',
-  marginBottom: '3vh'
-}
-
-const fontStyle = {
-    position: "absolute",
-    color: "#fff",
-    fontSize: "20px",
-    fontWeight:700,
-    margin: "5px",
-    marginLeft: "20px"
-}
-
-const smallFontStyle = {
-  color: "#fff",
-  textAlign: "right",
-  fontSize: "13px",
-  fontWeight: 300,
-  marginRight: "20px",
-  margin: "10px"
-}
-
-const cardWrapper = {
-  display: 'flex',
-  flexDirection: 'column',
-  justifyContent: 'center',
-  alignItems: 'center'
-}
-
-const modalStyle = {
+const Container = {
+  backgroundColor: '#000',
   position: 'absolute',
+  zIndex: '-1',
+  top: 0,
+  left: 0,
+  width: '100%',
+  height: '100%',
   display: 'flex',
-  top: '50%',
-  left: '50%',
-  width: '70vw',
-  maxWidth: 330,
-  height: "60vh",
-  transform: 'translate(-50%, -50%)',
-  borderRadius: 3,
-  bgcolor: 'background.paper',
-  boxShadow: 24,
+  justifyContent: 'center',
   alignItems: 'center',
-  flexDirection: 'column'
-};
+}
 
 const centerScreen = {
+  width: '70vw',
+  height: '50vh',
+  background: '#000000',
+  color: '#f5f5f5',
   display: 'flex',
   justifyContent: 'center',
   alignItems: 'center',
-  textAlign: 'center',
-  minHeight: '100vh',
+  marginLeft: '10%',
+  marginRight: '10%',
+  paddingTop: '15%',
+  paddingBottom: '10%',
+  flexDirection: 'column',
 }
 
 const button = {
@@ -101,205 +45,290 @@ const button = {
   backgroundColor: '#5f7182',
   color: '#fafafa',
   height: '50px',
-  margin: '30px 0 0 30px'
+  marginLeft: '45%'
 }
 
-const formTitle = {
-  height: '30px',
-  width: '426px',
-  border: 'none',
-  margin: '15px 25px 15px 15px',
-  textAlign: 'center',
+const exCenter = {
+  display: 'flex',
+  width: '50vw',
+  height: '40vh',
+  justifyContent: 'center',
+  alignItems: 'center',
+  margin: 'auto',
 }
 
-const formInput = {
-  boxSizing: 'border-box',
-  height: '39px',
-  width: '426px',
-  border: '1px solid #E6E6E6',
-  borderRadius: '10px',
-  margin: '15px 15px 15px 15px',
-  textIndent: '20px',
-}
-
-const useStyles = makeStyles({
-  root: {
-    borderWidth: "20px",
-    borderColor: "rgba(0,0,0,0.2)",
-    borderStyle: "solid"
-  }
-});
-
-function Cognitivetest () {
-  const classes = useStyles();
+const Cognitivetest2 = () => {
   const { userProfile, cognitiveTest, setCognitiveTest } = useMainContext();
+  const [userAuth, setUserAuth] = useState(false);
 
-  const [testListATrials, setTestListATrials] = useState(5);
-  const [firstPage, setFirstPage] = useState(true);
-  const [startTest, setStartTest] = useState(false);
-  const [tests, setTests] = useState({});
-  const [wordsArr, setWordsArr] = useState(['Asia', 'Sun', 'Key', 'Cat', 'Travel', 'Winter', '' ]);
-  const [word, setWord] = useState();
-  const [nav, setNav] = useState(false);
-  const [formState, setFormState] = useState(false);
-  const { register, handleSubmit, reset } = useForm();
-  const [data, setData] = useState([]);
-  const [val, setVal] = useState();
-  const [countModal, setCountModal] = useState(0);
-  const [finalComm, setFinalComm] = useState(false);
+  const [keyClicked, SetKeyClicked] = useState('')
+  const [pages, setPages] = useState( null )
+
+  const [TestIntro, setTestIntro] = useState(false)
+  const [startTest2, setStartTest2] = useState(false)
+  const [displayNum, setDisplayNum] = useState(true)
+  const [entry, setEntry] = useState('')
+  const [trial, setTrial] = useState(false)
+  const [displayCross, setDisplayCross] = useState(false)
+  const [test2End, setTest2End] = useState(false);
+
   const navigate = useNavigate();
 
+  const [loaded, setLoaded] = useState(false);
+  const [imgLoaded, setImgLoaded] = useState(false);
 
-  const handleRegistration = (data) => { setData(data); handleWordEnter(); reset(); }
+  const [trialEntry, setTrialEntry] = useState(12)
+  let trialCount = 12;
 
- useEffect(() => {
-   if (startTest) {
-    for (let i = 0; i < wordsArr.length; i++) {
-      (function(ind) {
-          setTimeout(function(){
-              setWord(wordsArr[ind]);
-          }, 1000 + (2000 * ind));
-      })(i);
+  const [isMounted, setIsMounted] = useState(false);
+  useEffect(() => {
+    setIsMounted(true);
+  }, []);
+
+  useEffect(() => {
+    return () => {
+      setIsMounted(false);
     }
-   }
- }, [startTest])
+  }, []);
 
+  useEffect(() => {
+    if (TestIntro) {
+      setTimeout(() => {
+        setImgLoaded(true)
+      }, 3000)
+      }
+  }, [TestIntro, loaded])
 
-// console.log('We are in stage 2 ...')
+  useEffect(() => {
+    if (userProfile) {
+      setUserAuth(true);
+    } else {
+      navigate('/login')
+    }
+   }, [userProfile])
 
-if (word === wordsArr[wordsArr.length - 1]) {
-  setNav(true);
-  setWord();
-}
+  //-----Listen to keyboard events
 
+  useEffect(() => {
+    document.addEventListener('keydown', detectSpaceBar, true)
+  }, [])
 
-const handleStartTest = (e) => {
-
-}
-
-const handleButtonClick = (e) => {
-  e.preventDefault();
-  setFormState(true);
-  setNav(false);
-}
-
-const toggleFormState = () => {
-  setFormState(true);
-}
-
-const handleWordEnter = () => {
-  setVal();
-  const count = (countModal + 1);
-  setCountModal(count);
-  setFormState(false);
-  if (count < wordsArr.length - 1) {
-    setTimeout(toggleFormState, 300);
+  const detectSpaceBar = (e) => {
+    SetKeyClicked(e.key);
   }
-};
 
-// useEffect(() => {
-//   console.log('Axios post data :', data)
-//   const param = {id: userProfile._id, word: data};
-//   axios.post('/api/cognitivetest', param)
-//   .then(() => {
-//     console.log(res.data)
-//   })
-// }, [data])
+  useEffect(() => {
+    if (keyClicked === 'Enter' && test2End) {
+      navigate('/cognitivetest3');
+    }
+  }, [keyClicked])
 
-useEffect(() => {
-  if (countModal === wordsArr.length - 1) {
-    const trialTimes = testListATrials - 1;
-    setTestListATrials(trialTimes);
-  }
-}, [countModal])
+  //-------------------------------------------
 
-useEffect(() => {
-  if (testListATrials > 0 && testListATrials < 5) {
-    console.log('How many tests left after test 2? : :', testListATrials)
-    setFinalComm(true);
-  }
-}, [testListATrials])
+  useEffect(() => {
+    setPages( document.querySelectorAll(".page") );
+  }, [])
 
+    let translateAmount = 100;
+    let translate = 0;
 
-useEffect(()=> {
-  if (countModal === wordsArr.length - 1 ) {
-    const modalCountdown = testListATrials - 1;
-    console.log('testListATrials before end :', testListATrials)
-    modalCountdown > 0 ? setTestListATrials(modalCountdown) : null;
-  }
-}, [countModal])
+    const slideNext = () => {
+      setTestIntro(true)
+      translate -= translateAmount ;
+      pages.forEach(
+        pages => (
+          pages.style.transform = `translateX(${translate}%)`)
+      );
+    }
 
-  const buttonEntry = countModal < wordsArr.length - 2 ? 'Enter' : 'Submit';
-  const wordCount = countModal + 1;
+    const slidePrev = () => {
+      translate += translateAmount;
+      pages.forEach(
+        pages => (pages.style.transform = `translateX(${translate}%)`)
+      );
+    }
+
+    const TestStart = () => {
+      setTestIntro(false);
+      setStartTest2(true);
+
+      setTimeout(() => {
+        setDisplayNum(false)
+        setDisplayCross(true)
+        setTrial(true)
+      }, 30000)
+
+      setTimeout(() => {
+        if (trialCount > 1) {
+          setDisplayCross(false)
+          setStartTest2(false);
+          setDisplayNum(true)
+          // timeInterval += 10000;
+          trialCount--;
+          TestStart()
+        } else {
+          let complete = {_id: userProfile._id, test2Completion: true};
+          axios.post('/api/user', complete)
+          .then((result) => {
+            console.log('success post test 2 completion', result)
+            setStartTest2(false)
+            setTest2End(true)
+          })
+        }
+      }, 60000)
+    }
+
+    useEffect(()=> {
+      let param;
+      let entryVal;
+      if (displayCross) {
+        entry === '' ? (entryVal = {}) : entryVal = entry;
+        if (trialEntry === 12) {
+          param = {id: userProfile._id, Test2_Entry_1: entryVal, };
+        } else if (trialEntry === 11) {
+          param = {id: userProfile._id, Test2_Entry_2: entryVal, };
+        } else if (trialEntry === 10) {
+          param = {id: userProfile._id,  Test2_Entry_3: entryVal, };
+        } else if (trialEntry === 9) {
+          param = {id: userProfile._id,  Test2_Entry_4: entryVal, };
+        } else if (trialEntry === 8) {
+          param = {id: userProfile._id,  Test2_Entry_5: entryVal, };
+        } else if (trialEntry === 7) {
+          param = {id: userProfile._id, Test2_Entry_6: entryVal, };
+        } else if (trialEntry === 6) {
+          param = {id: userProfile._id,  Test2_Entry_7: entryVal, };
+        } else if (trialEntry === 5) {
+          param = {id: userProfile._id,  Test2_Entry_8: entryVal, };
+        } else if (trialEntry === 4) {
+          param = {id: userProfile._id, Test2_Entry_9: entryVal, };
+        } else if (trialEntry === 3) {
+          param = {id: userProfile._id,  Test2_Entry_10: entryVal, };
+        } else if (trialEntry === 2) {
+          param = {id: userProfile._id,  Test2_Entry_11: entryVal, };
+        } else if (trialEntry === 1) {
+          param = {id: userProfile._id,  Test2_Entry_12: entryVal, };
+        }
+        setTrialEntry(trialEntry - 1);
+        axios.post('/api/cognitivetest', param )
+        .then((res) => {
+          console.log(res.data)
+          setEntry('')
+        })
+      }
+    }, [displayCross])
+
+    const handleEntry = (e) => {
+      e.preventDefault();
+      setEntry(e.target.value);
+    }
+
+    const handleStartTest3 = (e) => {
+      e.preventDefault();
+      navigate('/cognitivetest3');
+    }
+
   return (
     <>
-      <div className='centerScreen' style={centerScreen}>
+    <div className="container">
+      <div className="pages">
+        <>
+        {userAuth ?
+        <div id="page" className="page one">
+        <NavigationBar />
+          <div style={centerScreen} >
+            <h1 style={{ color: '#e67373', fontSize: 50, }} >TEST 2.</h1>
+            <div style={{ fontSize: 35, margin: '8% 0', fontWeight: 700, lineHeight: '1.6' }}>
+            You will be shown a sequence of numbers on the screen. Using your dominant hand (i.e., the hand you use to write with), type the sequence as fast and accurately as possible. Keep typing the sequence over and over until the numbers disappear. <br /> [Click next to continue]
+            </div>
+            <div>
+              <button onClick={slideNext} >Next</button>
+            </div>
+          </div>
+        </div>
+        : null}
 
-          <div style={{marginBottom: '350px', display: 'inline-block',}}  >
-            {firstPage ?
-                <>
-                  <Typography style={{fontSize: 60, marginBottom: 20, fontWeight: 700, display: 'inline-block', }}>
-                  Test 2: Finger Tapping Task.
-                  </Typography>
-                  <Typography style={{fontSize: 40, marginBottom: 20, color: '#be8484', fontWeight: 700, display: 'inline-block', }}>
-                  You will be shown a sequence of numbers on the screen. Using your dominant hand (i.e. the hand you use to write with), type the sequence as fast and accurately as possible. Keep typing the sequence over and over until the numbers disappear. [Press enter or click start to continue]
-                  </Typography>
-                  <Button variant="outlined" style={button} onClick={handleStartTest}  type='submit' >
-                  Start
-                  </Button>
-                </>
-                :
-              null}
-          </div>
-          <Typography style={{ fontSize: 40, fontWeight: 700, marginBottom: '300px', textAlign: 'center', fontFamily:'Courgette' }} >{word}</Typography>
-          {nav ?
+        <div id="page" className="page two">
+          <NavigationBar />
+          {TestIntro && !imgLoaded ?
           <>
-          <div style={{display: 'inline-block', marginBottom: '300px'}} >
-              <div style={{ fontSize: 40, fontWeight: 700}}>Click on button to enter as many words as you can remember...</div>
-              <Button variant="outlined" style={button} onClick={handleButtonClick} type='submit' >
-              Go..
-              </Button>
-          </div>
+              {/* <div className="loader" id="loader"></div> */}
+              <div className="loader" id="loader2"></div>
+              <span className="loadingText">LOADING...</span><br></br>
           </>
+          : null}
+          {TestIntro && imgLoaded ?
+          <div style={centerScreen} >
+            <div id='description' style={exCenter} >
+              <img
+              style= {loaded ? { width: '100%', height: '100%', marginBottom: '10vh', } : { display: 'none' } }
+              src="https://res.cloudinary.com/entazesoftware/image/upload/v1664984224/CogniticeTest-Keyboard_tlpkg2.png"
+              onLoad={() => setLoaded(true)}
+              />
+            </div>
+            <div style={{ fontSize: 35, fontWeight: 50, marginBottom: '6vh' }}>
+              Click previous to re-read instructions or click start to begin test.
+            </div>
+            <div style={{ display: "inline-block", margin: "4px" }} >
+              <button style={{ margin: "8px", background: "#000" }} onClick={slidePrev} >Previous</button>
+              <button style={{ margin: "8px", background: "#7e1515" }} onClick={TestStart} >Start</button>
+            </div>
+          </div>
           :
-          null }
-          {formState ?
+          null
+          }
+          {startTest2 ?
+          <div style={centerScreen} >
+            <div style={exCenter} >
+              {displayNum ?
+                <div style={{ display: 'flex', flexDirection: 'column', width: '100%', justifyContent: 'center', alignItems: 'center', height: '100%', }} >
+                  <h1 style={{ fontSize: 45, color: '#fff', paddingBottom: '90px', }} >
+                    41232
+                  </h1>
+                  <Box component="form" noValidate  sx={{ mt: 3, height: '35px', width: '100%' }}>
+                    <Grid item xs={12}>
+                      <TextField
+                        required
+                        fullWidth
+                        id="numberentry"
+                        name="numberentry"
+                        type="password"
+                        value={entry}
+                        style={{ width: '100%', height: '100%', background: '#fff', borderRadius: '3%' }}
+                        onChange={handleEntry}
+                        autoFocus
+                        autoComplete="off"
+                      />
+                    </Grid>
+                  </Box>
+                </div>
+              : null}
+              {displayCross ?
+                <img style={ {width:'190px', height: '140px' } } src="https://res.cloudinary.com/entazesoftware/image/upload/v1662136021/CognitiveApp/X-logo_mnf03d.jpg" />
+              : null}
+            </div>
+          </div>
+          : null}
+          {test2End ?
           <>
-            <div style={{ fontSize: 40, fontWeight: 700}}>
-              <Dialog open PaperProps={{ classes: { root: classes.root } }}>
-                <DialogContent>
-                  <div style={{display: 'inline-block'}} >
-                    <form onSubmit={handleSubmit(handleRegistration)}>
-                      {countModal < wordsArr.length - 2 ?
-                        <>
-                         <div style={formTitle}>Enter word {wordCount}..</div>
-                         <input name="name" {...register("word")} placeholder="Word.." style={formInput} />
-                        </>
-                        :
-                      null}
-                      <DialogActions>
-                        <Button variant="outlined" style={button} onClick={handleSubmit} type='submit' >
-                        {buttonEntry}
-                        </Button>
-                      </DialogActions>
-
-                    </form>
-                  </div>
-                </DialogContent>
-              </Dialog>
+            <div style={centerScreen} >
+            <div style={{ fontSize: 35, fontWeight: 700, display: 'flex', padding: '0px 40px 120px 40px', lineHeight: '1.6', }} >
+              Test 2 - Finger Tapping Task Completed.  <br /> [Click start or press ENTER to begin Test 3]
+            </div>
+            <div>
+              <button onClick={handleStartTest3}>Start</button>
+            </div>
             </div>
           </>
           :
           null}
-          {finalComm?
-           <Typography style={{fontSize: 40, fontWeight: 700, fontFamily:'Courgette'  }}>
-           Thank you ❤️!
-           </Typography>
-           :
-          null}
+        </div>
+        </>
       </div>
+    </div>
     </>
   )
 }
 
-export default Cognitivetest
+export default Cognitivetest2
+
+
